fix: invalid function prototype
the function prototype uses a 16bit wide integer instead of 8

+++ src/sync/match_config.c
@@ -1,7 +1,7 @@
 #include "header.h"
 #include "head_backup.h"
 
-int match_config(uint8_t buf_i, void *buf_arr[], Bool backup_data_arr[NUM_DATA_BACKUP], struct filename_type *filename_data)
+int match_config(uint16_t buf_i, void *buf_arr[], Bool backup_data_arr[NUM_DATA_BACKUP], struct filename_type *filename_data)
 {
 	Bool data_backed_up = False;
